--- cloud/pkg/admissioncontroller/constants/default.go
@@ -1,9 +1,0 @@
-package constants
-
-const (
-	WebhookNameDeviceModel = "deny-unwanted-devicemodel.k8s.io"
-	Port                   = 8899
-	NamespaceName          = "default"
-	ServiceName            = "webhook-service"
-	SecretName             = "webhook-secret"
-)
